user-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
pymysql==1.1.0
requests==2.31.0
python-jose==3.3.0
python-multipart==0.0.6 